refactor(utils): extract chrome helpers
